requirements.txt
pandas~=3.0.1
scikit-learn~=1.8.0
pyyaml~=6.0.3
feature-engine~=1.9.4
joblib~=1.5.3
streamlit~=1.55.0